+++ oshi-core/src/main/java/oshi/software/common/AbstractNetworkParams.java
@@ -1,5 +1,24 @@
+/**
+ * Oshi (https://github.com/dblock/oshi)
+ *
+ * Copyright (c) 2010 - 2017 The Oshi Project Team
+ *
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Maintainers:
+ * dblock[at]dblock[dot]org
+ * widdis[at]gmail[dot]com
+ * enrico.bianchi[at]gmail[dot]com
+ *
+ * Contributors:
+ * https://github.com/dblock/oshi/graphs/contributors
+ */
 package oshi.software.common;
 
+import com.sun.jna.ptr.PointerByReference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -8,6 +27,7 @@ import java.net.UnknownHostException;
 import java.util.ArrayList;
 import java.util.List;
 
+import oshi.jna.platform.unix.LibC;
 import oshi.software.os.NetworkParams;
 import oshi.util.FileUtil;
 
@@ -20,6 +40,41 @@ public abstract class AbstractNetworkParams implements NetworkParams {
 
     private static final Logger LOG = LoggerFactory.getLogger(AbstractNetworkParams.class);
     private static final String NAMESERVER = "nameserver";
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String getDomainName() {
+        LibC.Addrinfo hint = new LibC.Addrinfo();
+        hint.ai_flags = LibC.AI_CANONNAME;
+        String hostname = null;
+        try {
+            hostname = InetAddress.getLocalHost().getHostName();
+        } catch (UnknownHostException e) {
+            LOG.error("Unknown host exception when getting address of local host: " + e);
+            return "";
+        }
+        PointerByReference ptr = new PointerByReference();
+        doGetaddrinfo(hint, hostname, ptr);
+        LibC.Addrinfo info = new LibC.Addrinfo(ptr.getValue());
+        String canonname = info.ai_canonname;
+        doFreeaddrinfo(ptr);
+        int dot = canonname.indexOf('.');
+        if (dot == -1) {
+            return "";
+        } else {
+            return canonname.substring(dot + 1);
+        }
+    }
+
+    protected void doFreeaddrinfo(PointerByReference ptr) {
+        LibC.INSTANCE.freeaddrinfo(ptr.getValue());
+    }
+
+    protected void doGetaddrinfo(LibC.Addrinfo hint, String hostname, PointerByReference ptr) {
+        LibC.INSTANCE.getaddrinfo(hostname, null, hint, ptr);
+    }
 
     /**
      * {@inheritDoc}
@@ -62,10 +117,10 @@ public abstract class AbstractNetworkParams implements NetworkParams {
         return servers.toArray(new String[servers.size()]);
     }
 
-    static protected String searchGateway(List<String> lines){
-        for(String line: lines){
+    static protected String searchGateway(List<String> lines) {
+        for (String line : lines) {
             String leftTrimmed = line.replaceFirst("^[ \t]+", "");
-            if(leftTrimmed.startsWith("gateway:")){
+            if (leftTrimmed.startsWith("gateway:")) {
                 return leftTrimmed.split("[ \t]", 2)[1];
             }
         }